+++ src/DRSlib/multiprocessing.py
@@ -6,11 +6,36 @@ For when even the most basic parallel processing is
 better than nothing
 """
 import multiprocessing
+from multiprocessing.pool import Pool
 from typing import Any, Callable, List, Tuple
 
 
+class NoDaemonProcess(multiprocessing.Process):
+    @property
+    def daemon(self):
+        return False
+
+    @daemon.setter
+    def daemon(self, value):
+        pass
+
+
+class NoDaemonContext(type(multiprocessing.get_context())):
+    Process = NoDaemonProcess
+
+
+# We sub-class multiprocessing.pool.Pool instead of multiprocessing.Pool
+# because the latter is only a wrapper function, not a proper class.
+class NestablePool(Pool):
+    def __init__(self, *args, **kwargs):
+        kwargs["context"] = NoDaemonContext()
+        super(NestablePool, self).__init__(*args, **kwargs)
+
+
 class SimpleMultiProcessing:
-    """Parallel processing made simpler"""
+    """Parallel processing made simpler
+    (!) Since 0.8.0 SimpleMultiProcessing allows for nested multiprocessing. Use carefully !
+    """
 
     @staticmethod
     def apply_kwargs(
@@ -31,7 +56,7 @@ class SimpleMultiProcessing:
         user_function_and_arguments = [
             (user_function, (), _args) for _args in arguments
         ]
-        with multiprocessing.Pool(parallel_instances) as pool:
+        with NestablePool(parallel_instances) as pool:
             return list(
                 pool.imap(
                     SimpleMultiProcessing.apply_kwargs, user_function_and_arguments